Add missing rootElement option to types.d.ts
The lists a `rootElement` option, and that seems to exist in the types.ts file, but it is not listed in `index.d.ts`

This causes errors when trying to pass the `rootElement` option to the `EmojiButton`-costructor, even though there seems to be code that looks for that option in the constructor. The error messages look like this:

```
      TS2345: Argument of type '{ rootElement: any; }' is not assignable to parameter of type 'Options'.
  Object literal may only specify known properties, and 'rootElement' does not exist in type 'Options'.
```

This patch adds that missing `rootElement` option to the type definition.